[factory] refactore workingsetstore
